+++ src/Opulence/Debug/Exceptions/Handlers/IExceptionHandler.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2015 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Debug\Exceptions\Handlers;
+
+use Throwable;
+
+/**
+ * Defines the interface for exception handlers to implement
+ */
+interface IExceptionHandler
+{
+    /**
+     * Handles an exception
+     *
+     * @param Throwable $ex The exception to handle
+     */
+    public function handle($ex);
+
+    /**
+     * Registers the handler with PHP
+     */
+    public function register();
+}