+++ src/Opulence/Framework/Debug/Exceptions/Handlers/Console/IConsoleExceptionRenderer.php
@@ -6,10 +6,10 @@
  * @copyright Copyright (C) 2015 David Young
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
-namespace Opulence\Framework\Exceptions\Console;
+namespace Opulence\Framework\Debug\Exceptions\Handlers\Console;
 
 use Opulence\Console\Responses\IResponse;
-use Opulence\Exceptions\IExceptionRenderer;
+use Opulence\Debug\Exceptions\Handlers\IExceptionRenderer;
 
 /**
  * Defines the interface for console exception renderers to implement