+++ src/Opulence/Framework/Testing/PhpUnit/Http/ApplicationTestCase.php
@@ -13,8 +13,8 @@ namespace Opulence\Framework\Testing\PhpUnit\Http;
 
 use LogicException;
 use Opulence\Applications\Environments\Environment;
-use Opulence\Exceptions\ExceptionHandler;
-use Opulence\Framework\Exceptions\Http\IHttpExceptionRenderer;
+use Opulence\Debug\Exceptions\Handlers\IExceptionHandler;
+use Opulence\Framework\Debug\Exceptions\Handlers\Http\IHttpExceptionRenderer;
 use Opulence\Framework\Http\Kernel;
 use Opulence\Framework\Testing\PhpUnit\ApplicationTestCase as BaseApplicationTestCase;
 use Opulence\Http\Requests\Request;
@@ -254,7 +254,7 @@ abstract class ApplicationTestCase extends BaseApplicationTestCase
         $this->setApplicationAndIocContainer();
         $this->application->getEnvironment()->setName(Environment::TESTING);
         $this->application->start();
-        $this->container->bind(ExceptionHandler::class, $this->getExceptionHandler());
+        $this->container->bind(IExceptionHandler::class, $this->getExceptionHandler());
         $this->container->bind(IHttpExceptionRenderer::class, $this->getExceptionRenderer());
         $this->router = $this->container->makeShared(Router::class);
         $this->kernel = $this->container->makeShared(Kernel::class);